Refactor unit tests (#13)
* fixed find for missing args

* all unit tests passing, refactored find, unit test exceptions

* added missing argument exception

* updated cat unit tests 100% coverage

* deleted redundant command classes

* moved ShellVisitor to shell package

* readded shell visitor

* added missing argument exception cases

* head missing argument exception fix

* all tests EXCEPT stdin

* head tail super class

* Fixed missing argument exception head tail

* I promise this is the last missing argument fix for head tail

* I lied. Missing argument exception fix for grep, sort, uniq

* Fixed head test

* added standard input for head test

* tail unit tests

* uniq bug fix + uniq unit tests 100% coverage

* cat test assert equals parameter order fix

* added sort unit test 100% coverage

* sort bug fix + file not found exception fix

* refactored grep unit test

* grep correct exceptions

* Added missing exceptions for cut

* find correct exceptions

* cut unit tests 95% coverage

* cut unit test 100% coverage

* grep 95% 100%

* grep unit test 100% coverage

* head tail super class 100% coverage

* cd 100% branch coverage

* ls 100% branch coverage

* application factory 100% coverage

* substitution unit tests 100% coverage

* removed permanent test files dir

* added new exception for mkdir + changed formatting for exceptions

* Updated application factory unit test for new apps

* unsafe decorator unit tests refactored

* mkdir unit tests 100% coverage

* rm added missing exception

* rm unit tests 100% coverage

* sleep missing exception

* date logic error

* date unit tests 100% coverage

* sleep logic fix

* sleep units tests 93% coverage

* touch create multiple files fix

* Unit test clean up + touch unit tests

+++ src/main/java/uk/ac/ucl/shell/applications/Rm.java
@@ -1,6 +1,7 @@
 package uk.ac.ucl.shell.applications;
 
 import uk.ac.ucl.shell.Shell;
+import uk.ac.ucl.shell.exceptions.FileNotFoundException;
 import uk.ac.ucl.shell.exceptions.MissingArgumentsException;
 import uk.ac.ucl.shell.exceptions.TooManyArgumentsException;
 
@@ -9,7 +10,6 @@ import java.io.InputStream;
 import java.io.OutputStreamWriter;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.List;
 
 public class Rm implements Application {
@@ -23,6 +23,10 @@ public class Rm implements Application {
 
         String fileName = args.get(0);
         Path filePath = Shell.getCurrentDirectory().resolve(fileName);
-        Files.delete(filePath);
+        if (Files.exists(filePath)) {
+            Files.delete(filePath);
+        } else {
+            throw new FileNotFoundException("rm", filePath.toString());
+        }
     }
 }
